Restore TEE registry contract reads
